+++ extensions/youtube/src/main/java/app/revanced/extension/youtube/patches/theme/SeekbarColorPatch.java
@@ -18,6 +18,8 @@ public final class SeekbarColorPatch {
 
     private static final boolean SEEKBAR_CUSTOM_COLOR_ENABLED = Settings.SEEKBAR_CUSTOM_COLOR.get();
 
+    private static final boolean HIDE_SEEKBAR_THUMBNAIL_ENABLED = Settings.HIDE_SEEKBAR_THUMBNAIL.get();
+
     /**
      * Default color of the litho seekbar.
      * Differs slightly from the default custom seekbar color setting.
@@ -25,14 +27,19 @@ public final class SeekbarColorPatch {
     private static final int ORIGINAL_SEEKBAR_COLOR = 0xFFFF0000;
 
     /**
-     * Default colors of the gradient seekbar.
+     * Feed default colors of the gradient seekbar.
      */
-    private static final int[] ORIGINAL_SEEKBAR_GRADIENT_COLORS = { 0xFFFF0033, 0xFFFF2791 };
+    private static final int[] FEED_ORIGINAL_SEEKBAR_GRADIENT_COLORS = { 0xFFFF0033, 0xFFFF2791 };
 
     /**
-     * Default positions of the gradient seekbar.
+     * Feed default positions of the gradient seekbar.
      */
-    private static final float[] ORIGINAL_SEEKBAR_GRADIENT_POSITIONS = { 0.8f, 1.0f };
+    private static final float[] FEED_ORIGINAL_SEEKBAR_GRADIENT_POSITIONS = { 0.8f, 1.0f };
+
+    /**
+     * Empty seekbar gradient, if hide seekbar in feed is enabled.
+     */
+    private static final int[] HIDDEN_SEEKBAR_GRADIENT_COLORS = { 0x0, 0x0 };
 
     /**
      * Default YouTube seekbar color brightness.
@@ -41,15 +48,20 @@ public final class SeekbarColorPatch {
 
     /**
      * If {@link Settings#SEEKBAR_CUSTOM_COLOR} is enabled,
-     * this is the color value of {@link Settings#SEEKBAR_CUSTOM_COLOR_VALUE}.
+     * this is the color value of {@link Settings#SEEKBAR_CUSTOM_COLOR_PRIMARY}.
      * Otherwise this is {@link #ORIGINAL_SEEKBAR_COLOR}.
      */
-    private static int seekbarColor = ORIGINAL_SEEKBAR_COLOR;
+    private static int customSeekbarColor = ORIGINAL_SEEKBAR_COLOR;
 
     /**
      * Custom seekbar hue, saturation, and brightness values.
      */
     private static final float[] customSeekbarColorHSV = new float[3];
+
+    /**
+     * Custom seekbar color, used for linear gradient replacements.
+     */
+    private static final int[] customSeekbarColorGradient = new int[2];
 
     static {
         float[] hsv = new float[3];
@@ -63,26 +75,22 @@ public final class SeekbarColorPatch {
 
     private static void loadCustomSeekbarColor() {
         try {
-            seekbarColor = Color.parseColor(Settings.SEEKBAR_CUSTOM_COLOR_VALUE.get());
-            Color.colorToHSV(seekbarColor, customSeekbarColorHSV);
+            customSeekbarColor = Color.parseColor(Settings.SEEKBAR_CUSTOM_COLOR_PRIMARY.get());
+            Color.colorToHSV(customSeekbarColor, customSeekbarColorHSV);
+
+            customSeekbarColorGradient[0] = customSeekbarColor;
+            customSeekbarColorGradient[1] = Color.parseColor(Settings.SEEKBAR_CUSTOM_COLOR_ACCENT.get());
         } catch (Exception ex) {
             Utils.showToastShort(str("revanced_seekbar_custom_color_invalid"));
-            Settings.SEEKBAR_CUSTOM_COLOR_VALUE.resetToDefault();
+            Settings.SEEKBAR_CUSTOM_COLOR_PRIMARY.resetToDefault();
+            Settings.SEEKBAR_CUSTOM_COLOR_ACCENT.resetToDefault();
+
             loadCustomSeekbarColor();
         }
     }
 
     public static int getSeekbarColor() {
-        return seekbarColor;
-    }
-
-    /**
-     * Injection point
-     */
-    public static boolean playerSeekbarGradientEnabled(boolean original) {
-        if (SEEKBAR_CUSTOM_COLOR_ENABLED) return false;
-
-        return original;
+        return customSeekbarColor;
     }
 
     /**
@@ -126,7 +134,7 @@ public final class SeekbarColorPatch {
         // Even if the seekbar color xml value is changed to a completely different color (such as green),
         // a color filter still cannot be selectively applied when the drawable has more than 1 color.
         try {
-            String seekbarStyle = get9BitStyleIdentifier(seekbarColor);
+            String seekbarStyle = get9BitStyleIdentifier(customSeekbarColor);
             Logger.printDebug(() -> "Using splash seekbar style: " + seekbarStyle);
 
             final int styleIdentifierDefault = Utils.getResourceIdentifier(
@@ -148,6 +156,13 @@ public final class SeekbarColorPatch {
 
     /**
      * Injection point.
+     */
+    public static boolean showWatchHistoryProgressDrawable(boolean original) {
+        return !HIDE_SEEKBAR_THUMBNAIL_ENABLED && original;
+    }
+
+    /**
+     * Injection point.
      *
      * Overrides all Litho components that use the YouTube seekbar color.
      * Used only for the video thumbnails seekbar.
@@ -156,35 +171,61 @@ public final class SeekbarColorPatch {
      */
     public static int getLithoColor(int colorValue) {
         if (colorValue == ORIGINAL_SEEKBAR_COLOR) {
-            if (Settings.HIDE_SEEKBAR_THUMBNAIL.get()) {
-                return 0x00000000;
+            if (HIDE_SEEKBAR_THUMBNAIL_ENABLED) {
+                return 0x0;
             }
 
-            return getSeekbarColorValue(ORIGINAL_SEEKBAR_COLOR);
+            return customSeekbarColor;
         }
+
         return colorValue;
+    }
+
+    private static String colorArrayToHex(int[] colors) {
+        final int length = colors.length;
+        StringBuilder builder = new StringBuilder(length * 12);
+        builder.append("[");
+
+        int i = 0;
+        for (int color : colors) {
+            builder.append(String.format("#%X", color));
+            if (++i < length) {
+                builder.append(", ");
+            }
+        }
+
+        builder.append("]");
+        return builder.toString();
     }
 
     /**
      * Injection point.
      */
-    public static void setLinearGradient(int[] colors, float[] positions) {
-        final boolean hideSeekbar = Settings.HIDE_SEEKBAR_THUMBNAIL.get();
+    public static int[] getPlayerLinearGradient(int[] original) {
+        return SEEKBAR_CUSTOM_COLOR_ENABLED
+                ? customSeekbarColorGradient
+                : original;
+    }
 
-        if (SEEKBAR_CUSTOM_COLOR_ENABLED || hideSeekbar) {
+    /**
+     * Injection point.
+     */
+    public static int[] getLithoLinearGradient(int[] colors, float[] positions) {
+        if (SEEKBAR_CUSTOM_COLOR_ENABLED || HIDE_SEEKBAR_THUMBNAIL_ENABLED) {
             // Most litho usage of linear gradients is hooked here,
             // so must only change if the values are those for the seekbar.
-            if (Arrays.equals(ORIGINAL_SEEKBAR_GRADIENT_COLORS, colors)
-                    && Arrays.equals(ORIGINAL_SEEKBAR_GRADIENT_POSITIONS, positions)) {
-                Arrays.fill(colors, hideSeekbar
-                        ? 0x00000000
-                        : seekbarColor);
-                return;
+            if ((Arrays.equals(FEED_ORIGINAL_SEEKBAR_GRADIENT_COLORS, colors)
+                    && Arrays.equals(FEED_ORIGINAL_SEEKBAR_GRADIENT_POSITIONS, positions))) {
+                return HIDE_SEEKBAR_THUMBNAIL_ENABLED
+                        ? HIDDEN_SEEKBAR_GRADIENT_COLORS
+                        : customSeekbarColorGradient;
             }
 
-            Logger.printDebug(() -> "Ignoring gradient colors: " + Arrays.toString(colors)
+            Logger.printDebug(() -> "Ignoring gradient colors: " + colorArrayToHex(colors)
                     + " positions: " + Arrays.toString(positions));
         }
+
+        return colors;
     }
 
     /**
@@ -198,7 +239,7 @@ public final class SeekbarColorPatch {
         }
 
         return colorValue == ORIGINAL_SEEKBAR_COLOR
-                ? getSeekbarColorValue(ORIGINAL_SEEKBAR_COLOR)
+                ? customSeekbarColor
                 : colorValue;
     }
 
@@ -208,11 +249,9 @@ public final class SeekbarColorPatch {
      * Overrides color used for the video player seekbar.
      */
     public static int getVideoPlayerSeekbarColor(int originalColor) {
-        if (!SEEKBAR_CUSTOM_COLOR_ENABLED) {
-            return originalColor;
-        }
-
-        return getSeekbarColorValue(originalColor);
+        return SEEKBAR_CUSTOM_COLOR_ENABLED
+                ? getSeekbarColorValue(originalColor)
+                : originalColor;
     }
 
     /**
@@ -221,10 +260,6 @@ public final class SeekbarColorPatch {
      */
     private static int getSeekbarColorValue(int originalColor) {
         try {
-            if (!SEEKBAR_CUSTOM_COLOR_ENABLED || originalColor == seekbarColor) {
-                return originalColor; // nothing to do
-            }
-
             final int alphaDifference = Color.alpha(originalColor) - Color.alpha(ORIGINAL_SEEKBAR_COLOR);
 
             // The seekbar uses the same color but different brightness for different situations.
@@ -237,7 +272,7 @@ public final class SeekbarColorPatch {
             hsv[1] = customSeekbarColorHSV[1];
             hsv[2] = clamp(customSeekbarColorHSV[2] + brightnessDifference, 0, 1);
 
-            final int replacementAlpha = clamp(Color.alpha(seekbarColor) + alphaDifference, 0, 255);
+            final int replacementAlpha = clamp(Color.alpha(customSeekbarColor) + alphaDifference, 0, 255);
             final int replacementColor = Color.HSVToColor(replacementAlpha, hsv);
             Logger.printDebug(() -> String.format("Original color: #%08X  replacement color: #%08X",
                             originalColor, replacementColor));